one file per relational operator
